+++ chassembly/src/main/java/org/quiltmc/chasm/lang/api/ast/MapNode.java
@@ -8,6 +8,7 @@ import java.util.Map;
 import org.jetbrains.annotations.ApiStatus;
 import org.quiltmc.chasm.lang.api.eval.Evaluator;
 import org.quiltmc.chasm.lang.api.eval.Resolver;
+import org.quiltmc.chasm.lang.internal.render.RenderUtil;
 import org.quiltmc.chasm.lang.internal.render.Renderer;
 
 public class MapNode extends Node {
@@ -29,11 +30,7 @@ public class MapNode extends Node {
         for (int i = 0; i < list.size(); i++) {
             renderer.indent(builder, currentIndentationMultiplier);
             String key = list.get(i).getKey();
-            if (needsQuotes(key)) {
-                builder.append('"').append(key.replace("\\", "\\\\").replace("\"", "\\\"")).append('"');
-            } else {
-                builder.append(key);
-            }
+            builder.append(RenderUtil.quotifyIdentifierIfNeeded(key, '"'));
             builder.append(": ");
             list.get(i).getValue().render(renderer, builder, currentIndentationMultiplier + 1);
             if (i < entries.size() - 1 || renderer.hasTrailingCommas()) {
@@ -44,32 +41,6 @@ public class MapNode extends Node {
             renderer.indent(builder, currentIndentationMultiplier - 1);
         }
         builder.append('}');
-    }
-
-    private static boolean needsQuotes(String key) {
-        if (key.isEmpty()) {
-            return true;
-        }
-
-        if (!isValidIdentifierStartChar(key.charAt(0))) {
-            return true;
-        }
-
-        for (int i = 1; i < key.length(); i++) {
-            if (!isValidIdentifierChar(key.charAt(i))) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private static boolean isValidIdentifierStartChar(char c) {
-        return (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z') || c == '_';
-    }
-
-    private static boolean isValidIdentifierChar(char c) {
-        return isValidIdentifierStartChar(c) || (c >= '0' && c <= '9');
     }
 
     @Override
