+++ chassembly/src/main/java/org/quiltmc/chasm/lang/api/ast/ValueNode.java
@@ -3,6 +3,7 @@ package org.quiltmc.chasm.lang.api.ast;
 import org.jetbrains.annotations.ApiStatus;
 import org.quiltmc.chasm.lang.api.eval.Evaluator;
 import org.quiltmc.chasm.lang.api.eval.Resolver;
+import org.quiltmc.chasm.lang.internal.render.RenderUtil;
 import org.quiltmc.chasm.lang.internal.render.Renderer;
 
 public abstract class ValueNode<T> extends Node {
@@ -23,7 +24,7 @@ public abstract class ValueNode<T> extends Node {
     @Override
     public void render(Renderer renderer, StringBuilder builder, int currentIndentationMultiplier) {
         if (value instanceof String) {
-            builder.append('"').append(((String) value).replace("\\", "\\\\").replace("\"", "\\\"")).append('"');
+            builder.append(RenderUtil.quotify((String) value, '"'));
         } else {
             builder.append(value);
         }
