+++ chassembly/src/main/java/org/quiltmc/chasm/lang/internal/parse/Parser.jj
@@ -29,7 +29,7 @@ import org.quiltmc.chasm.lang.api.ast.UnaryNode;
 import org.quiltmc.chasm.lang.internal.parse.SourceSpan;
 
 public class Parser {
-    private static String extractStringOrCharLiteral(String image) {
+    private static String extractTextualLiteral(String image) {
         StringBuilder sb = new StringBuilder(image.length() - 2);
         for (int i = 1, e = image.length() - 1; i < e; i++) {
             char c = image.charAt(i);
@@ -85,10 +85,10 @@ Node literalExpression():
         { n = new FloatNode(Double.parseDouble(t.getImage())); }
         |
         t = <StringLiteral>
-        { n = new StringNode(extractStringOrCharLiteral(t.getImage())); }
+        { n = new StringNode(extractTextualLiteral(t.getImage())); }
         |
         t = <CharLiteral>
-        { n = new IntegerNode((long) extractStringOrCharLiteral(t.getImage()).charAt(0)); }
+        { n = new IntegerNode((long) extractTextualLiteral(t.getImage()).charAt(0)); }
     )
 
     {
@@ -109,9 +109,13 @@ Node referenceExpression():
         "$"
         { g = true; }
     )?
-    t = <Identifier>
+    ( t = <Identifier> | t = <ReferenceLiteral> )
     {
-        Node n = new ReferenceNode(t.getImage(), g);
+        String s = t.getImage();
+        if (s.startsWith("`")) {
+            s = extractTextualLiteral(s);
+        }
+        Node n = new ReferenceNode(s, g);
         SourceSpan span = SourceSpan.fromToken(start).join(SourceSpan.fromToken(lastConsumedToken));
         n.getMetadata().put(SourceSpan.class, span);
         return n;
@@ -175,7 +179,7 @@ Node mapExpression():
                 { key = t.getImage(); }
                 |
                 t = <StringLiteral>
-                { key = extractStringOrCharLiteral(t.getImage()); }
+                { key = extractTextualLiteral(t.getImage()); }
             )
             <Colon>
             e = expression()
@@ -188,7 +192,7 @@ Node mapExpression():
                     { key = t.getImage(); }
                     |
                     t = <StringLiteral>
-                    { key = extractStringOrCharLiteral(t.getImage()); }
+                    { key = extractTextualLiteral(t.getImage()); }
                 )
                 <Colon>
                 e = expression()
@@ -260,9 +264,13 @@ Node argumentExpression():
         |
         (
             <Dot>
-            t = <Identifier>
+            ( t = <Identifier> | t = <ReferenceLiteral> )
             {
-                e = new MemberNode(e, t.getImage());
+                String s = t.getImage();
+                if (s.startsWith("`")) {
+                    s = extractTextualLiteral(s);
+                }
+                e = new MemberNode(e, s);
                 SourceSpan span = SourceSpan.fromToken(start).join(SourceSpan.fromToken(lastConsumedToken));
                 e.getMetadata().put(SourceSpan.class, span);
             }
@@ -634,6 +642,10 @@ TOKEN: {
     <#Char: ~["'", "\\"] | "\\'" | "\\\\" >
     |
     <CharLiteral: "'" <Char> "'" >
+    |
+    <#ReferenceLiteralChar: ~["`", "\\"] | "\\`" | "\\\\" >
+    |
+    <ReferenceLiteral: "`" (<ReferenceLiteralChar>)* "`" >
 }
 
 // Identifier
