+++ chassembly/src/test/resources/results/literals/basic.chasm
@@ -12,4 +12,5 @@
     backslash: 92, 
     single_quote: 39, 
     "quoted key": 0, 
+    quoted_key_ref: 0, 
 }
